+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Version 1.3 (Status: DEVELOPMENT)
 * nafsdmctl: now has remove by domain and remove by ID functions
 * nafsdmctl: now has status/start/stop/restart functions for the webinterface
 * Slaves now report last connection time to master (nafsdmctl can show a table of which slaves has connected recently)
+* nafscli logviewer no longer throws a traceback message on exit (using CTRL+C)
 
 Version 1.2.5 (Status: RELEASED)
 * Added nafscli for the slave, a command-line interface